Delete firestore_logging_util.h, and move its contents into firestore_integration_test.h
