+++ firestore/integration_test_internal/src/firestore_integration_test.h
@@ -174,6 +174,27 @@ class TestEventListener {
   bool print_debug_info_ = false;
 };
 
+// A RAII wrapper that enables Firestore debug logging and then disables it
+// upon destruction.
+//
+// This is useful for enabling Firestore debug logging in a specific test.
+//
+// Example:
+// TEST(MyCoolTest, VerifyFirestoreDoesItsThing) {
+//   FirestoreDebugLogEnabler firestore_debug_log_enabler;
+//   ...
+// }
+class FirestoreDebugLogEnabler {
+ public:
+  FirestoreDebugLogEnabler() {
+    Firestore::set_log_level(LogLevel::kLogLevelDebug);
+  }
+
+  ~FirestoreDebugLogEnabler() {
+    Firestore::set_log_level(LogLevel::kLogLevelInfo);
+  }
+};
+
 // Base class for Firestore integration tests.
 // Note it keeps a cache of created Firestore instances, and is thread-unsafe.
 class FirestoreIntegrationTest : public testing::Test {
