+++ firestore/integration_test_internal/src/numeric_transforms_test.cc
@@ -21,7 +21,6 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "util/event_accumulator.h"
-#include "util/firestore_logging_util.h"
 
 namespace firebase {
 namespace firestore {
